+++ Assets/VRM10/Runtime/IO/Model.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 18a0f7c989f911547aa4b75abb583b4c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
