+++ Assets/VRM10/Runtime/IO/Model/MeshReader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2f4647a633dfc844eb1b8e7563addba2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
